Implement Children
This change implements children within the learning system. Children are associated with a family and have a set of skills associated with them. Each skill is a simple ID relating to a specific subsection of a skill description document which powers both the frontend and backend parts of this system.

Any family member can update the known skills of a child.

The family's children are returned when querying for the family. There is no way to query an individual child or to get the children explicitly because this functionality isn't needed by the frontend currently but a full RESTFUL interface would have these representations as well.