Implement child progress page
This chnage implements the child progress page. Progress is measured across a set of benchmarks which can be marked as having been completed or not. Each benchmark has an associated theme and age range. The display can be filtered due to the large number of benchmarks.

As part of this work the data format was enhanced to support sentinel benchmarks since otherwise there would be no way for the backend to know that the last benchmark in a series had been completed. The sentinel indicates completion of a series.