Implement Activities Backend
This change implements the backend for activities. An activity represents something that can be done with a child for entertainment or education. Each activity has a title, description, and a set of links to help further describe the activity.

Activities are associated with families rather than with users or children. When accepting an invitation to family, the activities from the previous family are moved over to the new family.

This change also fixes a bug in which error messages would be returned as an array of individual characters rather than as a string.