Implement OAuth For Mobile
Safari mobile, unfortunately, is unable to use the normal Google sign in button and so an alternative authentication mechanism is required. This change implements OAuth specifically for this scenario and makes it available only in narrow window scenarios.

Implementing this requires the use of sessions to store the authentication state of the mobile application. Authentication is refactored to enable this and share relevant code between the normal JWT-based authentication and the new OAuth-based authentication.

Frontend changes are likewise needed to support not sending a JWT when session-based authentication is used as well as to vary the view slightly when using a mobile browser. In general, mobile browsers can consume content but not edit it.